requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
gunicorn==21.2.0